Apply 4 suggestion(s) to 3 file(s)

+++ aruw-mcb-project/src/aruwsrc/control/auto-aim/auto_aim_launch_timer.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020-2022 Advanced Robotics at the University of Washington <robomstr@uw.edu>
+ * Copyright (c) 2022 Advanced Robotics at the University of Washington <robomstr@uw.edu>
  *
  * This file is part of aruw-mcb.
  *
@@ -69,7 +69,7 @@ AutoAimLaunchTimer::LaunchInclination AutoAimLaunchTimer::getCurrentLaunchInclin
     uint32_t projectedHitTime = now + this->agitatorTypicalDelayMicroseconds + timeOfFlightMicros;
 
     uint32_t nextPlateTransitTime = aimData.timestamp + aimData.targetHitTimeOffset;
-    int64_t projectedHitTimeAfterFirstWindow = int64_t(projectedHitTime) - nextPlateTransitTime;
+    int64_t projectedHitTimeAfterFirstWindow = int64_t(projectedHitTime) - int64_t(nextPlateTransitTime);
 
     int64_t offsetInFiringWindow = projectedHitTimeAfterFirstWindow % aimData.targetPulseInterval;
     if (offsetInFiringWindow < 0)
